Move the common constants into elasticai_api and add elasticai_api in the requirements of elasticdl.
