+++ elasticdl/python/common/constants.py
@@ -86,21 +86,3 @@ class ReaderType(object):
 
 class Initializer(object):
     UNIFORM = "uniform"
-
-
-class HorovodEnv(object):
-    RENDEZVOUS_ADDR = "HOROVOD_GLOO_RENDEZVOUS_ADDR"
-    RENDEZVOUS_PORT = "HOROVOD_GLOO_RENDEZVOUS_PORT"
-    RANK = "HOROVOD_RANK"
-    SIZE = "HOROVOD_SIZE"
-    CONTROLLER = "HOROVOD_CONTROLLER"
-    CPU_OPERATIONS = "HOROVOD_CPU_OPERATIONS"
-    HOSTNAME = "HOROVOD_HOSTNAME"
-    ELASTIC = "HOROVOD_ELASTIC"
-    GLOO_TIMEOUT_SECONDS = "HOROVOD_GLOO_TIMEOUT_SECONDS"
-
-
-class WorkerEnv(object):
-    MASTER_ADDR = "MASTER_ADDR"
-    WORKER_ID = "WORKER_ID"
-    WORKER_NUM = "WORKER_NUM"
