+++ elasticdl/python/master/pod_manager.py
@@ -22,12 +22,9 @@ from collections import Counter
 
 from kubernetes.client import V1EnvVar
 
+from elasticai_api.common.constants import WorkerEnv
 from elasticdl.python.common import k8s_client as k8s
-from elasticdl.python.common.constants import (
-    PodManagerStatus,
-    PodStatus,
-    WorkerEnv,
-)
+from elasticdl.python.common.constants import PodManagerStatus, PodStatus
 from elasticdl.python.common.k8s_client import PodType
 from elasticdl.python.common.log_utils import default_logger as logger
 from elasticdl.python.common.model_utils import get_dict_from_params_str
