+++ setup.py
@@ -15,6 +15,7 @@ from setuptools import find_packages, setup
 
 with open("elasticdl/requirements.txt") as f:
     required_deps = f.read().splitlines()
+required_deps.append("elasticai_api")
 required_deps.append("elasticdl_client")
 required_deps.append("elasticdl_preprocessing")
